Adopt komega and rename test
